fix wrongful configuration getter, core.py should now fetch set port properly
